+++ R/rc_tables.R
@@ -22,7 +22,7 @@
 #'
 #' This table contains the makkink correction factors for evapo-transpiration per month
 #'
-#' @format A data.table with 12 rows and 4 columns:
+#' @format A data.table with 13 rows and 4 columns:
 #' \describe{
 #'   \item{crop_makkink}{Makkink crop category}
 #'   \item{1}{Evapotranspiration correction factors for January}