+++ NEWS.md
@@ -1,3 +1,14 @@
+# rotsee UNRELEASED
+## added
+* examples for the functions `rc_calculate_bd`, `rc_calculate_bcof`, `rc_extend_crops`, `rc_extend_amendments`
+* unit tests for rc_visualize_plot
+* Vignettes to add overview of rotsee and a description of the helper functions
+
+## changed
+* removed intermediate columns dens.sand, dens.clay, and cf as output from `rc_calculate_bd`
+* set default of initialize in `parms` to FALSE
+* updated function name `rc_calculate_B_C_OF` to `rc_calculate_bcof`.
+
 # rotsee v0.2.2 2026-01-09
 ## added
 * test helper factories to aid with unit test creation
@@ -18,6 +29,7 @@
 ## changed
 * news of `rotsee v0.1.4`, to better reflect the updates of introduced parameters table
 * updated parameter table with new inputs
+
 
 # rotsee v0.2.0 2025-12-10